apps/gaslight/backend/requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
